requirements.txt
alembic == 1.8.1
sqlalchemy == 1.4.45
sqlalchemy-orm == 1.2.3
python-dotenv == 0.21.0
psycopg2-binary == 2.9.5